pyproject.toml
[tool.poetry]
name = "rag-fundamentals"
version = "0.1.0"
description = "Minimal end-to-end RAG slice: CLI + FAISS/BM25 retrieval + eval + safety."
authors = ["You <czuba.przemyslaw@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "=3.11.13"
openai = "^1.51.0"
tiktoken = "^0.7.0"
faiss-cpu = "^1.8.0.post1"
sentence-transformers = "^3.0.1"
rank-bm25 = "^0.2.2"
numpy = "^2.1.0"
jinja2 = "^3.1.4"
pydantic = "^2.8.2"
typer = "^0.12.5"
jsonlines = "^4.0.0"
rich = "^13.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
docbot = "src.cli:app"
